3 test cases for "if statements"
This is in preparation for piecewise function. Each "if statement" is a
Level-4 expression followed by a comma and a logical expression. The
omission of the "if" keyword is for the sake of simplicity of handwriting
input. I may decide to add support for this keyword in the syntax in the
future.